+++ src/Symfony/Component/Form/Extension/Core/Type/UrlType.php
@@ -39,11 +39,7 @@ class UrlType extends AbstractType
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'default_protocol' => static function (Options $options) {
-                trigger_deprecation('symfony/form', '7.1', 'Not configuring the "default_protocol" option when using the UrlType is deprecated. It will default to "null" in 8.0.');
-
-                return 'http';
-            },
+            'default_protocol' => null,
             'invalid_message' => 'Please enter a valid URL.',
         ]);